* real/main: (96 commits)
  ansible-plugin 3.2.2 upgrade (winrm non-admin fix)
  fix and tests for RSE-9
  RUN-1092: Update okhttp to 4.9.2 (#7825)
  RUN-1089: Upgrade to Liquibase 4.8.0 (#7831)
  now considering case with execution and schedule status diferent but at same time. Added tests.
  RCLOUD-561 - Add a location on login page where another login button can be inserted
  Make UpdateModeProjectService generate only one ExecutionStatusJob when having both disable/enable execution and schedule with the same time.
  Fix: tomcat docker test requires tag argument to script
  Upgrade grails to 5.1.9
  fix project validation for referenced executions
  improve query to get referenced executions
  Upgrade spring to 5.3.20
  Temporarily remove the scan for PR but keep the nightly scan
  Test slack notification
  Set correct gradle grails plugin version
  fix spring version setting. consolidate liquibase version. upgrade gson, jetty and grails.
  Minor refector
  Revert "RUN-1082: Add audit tracking for jobs and acls."
  Minor refactor
  Minor refactor
  ...

+++ rundeckapp/src/main/groovy/org/rundeck/app/api/model/SystemInfoModel.groovy
@@ -90,9 +90,9 @@ class SystemInfoModel {
         @Schema(name="memory", description="Memory Information")
         static class MemoryInfoModel {
             String unit
-            int max
-            int free
-            int total
+            long max
+            long free
+            long total
         }
         @Schema(name="scheduler", description="Scheduler Information")
         static class SchedulerInfoModel {
